feat(desktop): handle sem params on signin page
